Split out embeds for maintainability
